feat(router): add healthRouter to appRouter to provide health check endpoint
A new file `health-router/router.ts` was added to the `packages/trpc/server` directory. This file contains a `healthRouter` that provides a health check endpoint. The endpoint returns a JSON object with a `status` property set to `'ok'`.

In the `router.ts` file, the `healthRouter` was imported and added to the `appRouter` as a new route. This allows the health check endpoint to be accessible through the main application router.

The purpose of this change is to provide a simple health check endpoint that can be used to monitor the status of the server.